requirements.txt
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
python-dotenv>=0.19.0
requests>=2.28.0